last changes pak2/lab01 shit

+++ pak2/lab1/additional.h
@@ -9,8 +9,8 @@ typedef enum {
 }ReturnCode;
 
 typedef struct {
-    long numer;
-    long denumer;
+    long long numer;
+    long long denumer;
 
 }Fraction;
